requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
click==6.7
colorlog==3.1.0
flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
kazoo==2.4.0              # via pykafka
markupsafe==1.0           # via jinja2
py-zabbix==1.1.3
pykafka==2.6.0
pyyaml==3.12
six==1.11.0               # via kazoo, pykafka, structlog
structlog==17.2.0
tabulate==0.7.7           # via pykafka
werkzeug==0.12.2          # via flask
